Move FirRenderer class inside FirVisualizer to avoid ambiguity with existed FirRenderer
